#87 - refactor: 게시글 Controller 단일 책임 원칙 적용
